requirements/base.txt
Django==1.8
django-annoying==0.8.1
django-compressor==1.4
django-suit==0.2.13
sorl-thumbnail==12.2
bpython==0.14.1
redis==2.10.3
django-tinymce==1.5.3
pytz==2015.2
django-suit-redactor==0.0.2
Unipath==1.1
Pillow==2.8.1
mailchimp==2.0.9
psycopg2==2.6
django-markdown
Pygments
